Create change-file-extension
Given a list of filenames, we want to rename all the files with extension jpg to extension j. To do this, we would like to generate a new list called newfilenames, consisting of the new filenames. Fill in the blanks in the code using any of the methods you’ve learned thus far, like a for loop or a list comprehension.