Update convert-octal-file-permission-to-s
The Linux file system are split into three sets of three permission:red, write and execute for the owner, group and others. This three value is represented with the summing of each permission, with 4 corresponding to read, 2 to write and 1 to execute. Also it can be written with a string using the letters r,w, and x or - when permission is not granted