Create decimal_to_binary.py
This python script converts a decimal values to binary. 
Here are the algorithm I've followed:

Find the highest decimal value that is less than or equal to 214. This value is 128.
Set the bit in the 128 column to 1.
Subtract 128 from 214 to get a remainder of 86.
Find the highest decimal value that is less than or equal to 86. This value is 64.
Set the bit in the 64 column to 1.
Subtract 64 from 86 to get a remainder of 22.
Repeat the preceding pattern for each remainder value until the remainder is 0.